+++ ui/litellm-dashboard/src/app/(dashboard)/cost-tracking/components/how_it_works.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "../../../tests/test-utils";
+import { renderWithProviders } from "../../../../../tests/test-utils";
 import HowItWorks from "./how_it_works";
 
 vi.mock("@/app/(dashboard)/api-reference/components/CodeBlock", () => ({
