+++ ui/litellm-dashboard/src/app/(dashboard)/cost-tracking/components/pricing_calculator/multi_cost_results.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { renderWithProviders } from "../../../../tests/test-utils";
+import { renderWithProviders } from "../../../../../../tests/test-utils";
 import MultiCostResults from "./multi_cost_results";
 import type { MultiModelResult } from "./types";
 import type { CostEstimateResponse } from "../types";
