+++ ui/litellm-dashboard/src/app/(dashboard)/cost-tracking/components/use_margin_config.test.ts
@@ -18,7 +18,7 @@ vi.mock("./provider_display_helpers", () => ({
   }),
 }));
 
-vi.mock("../provider_info_helpers", () => ({
+vi.mock("@/components/provider_info_helpers", () => ({
   Providers: {
     OpenAI: "OpenAI",
     Anthropic: "Anthropic",
